프로젝트 업데이트 작업 - 7

+++ src/main/java/com/used/lux/dto/user/useraccount/UserAccountDto.java
@@ -13,7 +13,7 @@ public record UserAccountDto(
         String phoneNumber,
         int age,
         String gender,
-        int userPoint,
+        Long userPoint,
         UserGrade userGrade,
         RoleType role,
         String memo,
